generate cpp by pdll
