--- oneflow/ir/include/OneFlow/expr.pdll
@@ -1,1 +1,0 @@
-Pattern => erase op<>;
