+++ oneflow/ir/lib/OneFlow/PDLL/TestPDLL.pdll
@@ -1,0 +1,1 @@
+Pattern => erase op<>;
