+++ oneflow/ir/lib/OneFlow/PDLL/lit.local.cfg
@@ -1,0 +1,1 @@
+config.suffixes.remove('.pdll')
