+++ oneflow/ir/lib/OneFlow/Passes.cpp
@@ -41,7 +41,6 @@ limitations under the License.
 #include "mlir/IR/Attributes.h"
 #include "mlir/IR/OperationSupport.h"
 #include "mlir/IR/MLIRContext.h"
-#include "mlir/Parser/Parser.h"
 
 #include "mlir/Conversion/LinalgToLLVM/LinalgToLLVM.h"
 #include "mlir/Conversion/MemRefToLLVM/MemRefToLLVM.h"
@@ -728,7 +727,6 @@ bool IsSameDtype(mlir::OpResult cast_result, mlir::Value input) {
 }  // namespace mlir
 
 #include "OneFlow/OneFlowPatterns.cpp.inc"
-#include "OneFlow/TestPDLLPatterns.h.inc"
 
 namespace mlir {
 
@@ -1052,8 +1050,7 @@ void populatePreConvertInferenceOp(::mlir::RewritePatternSet& patterns) {
 }
 
 void populateConvertInferenceOp(::mlir::RewritePatternSet& patterns) {
-  // patterns.add<FuseConv2DBatchNormPattern>(patterns.getContext());
-  patterns.add<GeneratedPDLLPattern0>(patterns.getContext());
+  patterns.add<FuseConv2DBatchNormPattern>(patterns.getContext());
 }
 
 void populatePostConvertInferenceOp(::mlir::RewritePatternSet& patterns) {
